refactor interactive shell routing and test fixtures
Consolidate command routing into a dedicated resolve_cli_command package, move prompt-driven routing fixtures under tests/prompts, and simplify non-command fallback behavior to match the new two-path router design.
