+++ app/cli/interactive_shell/loop/dispatch.py
@@ -1,0 +1,247 @@
+"""Turn-dispatch logic for the interactive shell loop."""
+
+from __future__ import annotations
+
+import re
+import sys
+import threading
+from collections.abc import Callable
+
+from prompt_toolkit.key_binding import KeyBindings, merge_key_bindings
+from prompt_toolkit.key_binding.key_processor import KeyPressEvent
+from rich.console import Console
+
+from app.cli.interactive_shell.prompting import prompt_surface as _prompt_surface
+from app.cli.interactive_shell.routing import router as _router
+from app.cli.interactive_shell.runtime import HotReloadCoordinator, ReplSession
+from app.cli.interactive_shell.ui import render_banner
+from app.cli.interactive_shell.ui.choice_menu import repl_tty_interactive
+
+from .execution import execute_routed_turn
+from .state import _PROMPT_REFRESH_INTERVAL_S, ReplState
+
+render_submitted_prompt = _prompt_surface.render_submitted_prompt
+resolve_cli_command = _router.resolve_cli_command
+
+_INTERVENTION_CORRECTION_RE = re.compile(
+    r"("
+    r"no(?=[,.!?]|$)"
+    r"|nope\b"
+    r"|nvm\b"
+    r"|nevermind\b|never\s*mind\b"
+    r"|wrong\b"
+    r"|wait(?=[,.!?]|$)"
+    r"|stop(?=[,.!?]|$)"
+    r"|actually\b"
+    r"|scratch\s+that\b"
+    r"|instead(?=[,.!?]|$)"
+    r"|(?:let'?s\s+)?do\s+[^.\n]{1,60}\s+instead\b"
+    r"|try\s+[^.\n]{1,60}\s+instead\b"
+    r")",
+    re.IGNORECASE,
+)
+_CONFIRMATION_TOKENS: frozenset[str] = frozenset({"", "y", "yes", "n", "no"})
+_CANCEL_REQUEST_TOKENS: frozenset[str] = frozenset({"/cancel", "/stop", "/abort"})
+_EXCLUSIVE_STDIN_MENU_COMMANDS: frozenset[str] = frozenset(
+    {
+        "/history",
+        "/help",
+        "/integrations",
+        "/list",
+        "/mcp",
+        "/model",
+        "/template",
+        "/trust",
+        "/verbose",
+        "/?",
+    }
+)
+_EXCLUSIVE_STDIN_SUBCOMMANDS: frozenset[tuple[str, str]] = frozenset(
+    {
+        ("/integrations", "setup"),
+        ("/mcp", "connect"),
+    }
+)
+_WAIT_FOR_COMPLETION_COMMANDS: frozenset[str] = frozenset({"/exit", "/quit"})
+
+
+class DispatchCancelled(Exception):
+    """Raised when in-flight dispatch is cancelled during confirmation."""
+
+
+def looks_like_confirmation_answer(text: str | None) -> bool:
+    return (text or "").strip().lower() in _CONFIRMATION_TOKENS
+
+
+def looks_like_cancel_request(text: str | None) -> bool:
+    return (text or "").strip().lower() in _CANCEL_REQUEST_TOKENS
+
+
+def _looks_like_correction(text: str) -> bool:
+    stripped = text.lstrip()
+    if not stripped or stripped.startswith("```"):
+        return False
+    return _INTERVENTION_CORRECTION_RE.match(stripped[:80]) is not None
+
+
+def dispatch_should_show_spinner(text: str, session: ReplSession) -> bool:
+    return resolve_cli_command(text.strip(), session) is None
+
+
+def dispatch_needs_exclusive_stdin(text: str, session: ReplSession) -> bool:
+    if not repl_tty_interactive():
+        return False
+
+    t = text.strip()
+    if not t:
+        return False
+
+    command_decision = resolve_cli_command(t, session)
+    if command_decision is None:
+        return False
+    dispatch_text = command_decision.command_text or t
+
+    parts = dispatch_text.split()
+    if not parts:
+        return False
+    name = parts[0].lower()
+    args = [arg.lower() for arg in parts[1:]]
+
+    if name in _WAIT_FOR_COMPLETION_COMMANDS:
+        return True
+    if name in _EXCLUSIVE_STDIN_MENU_COMMANDS and not args:
+        return True
+    if name == "/tests" and not args:
+        return True
+    return bool(args and (name, args[0]) in _EXCLUSIVE_STDIN_SUBCOMMANDS)
+
+
+def dispatch_one_turn(
+    text: str,
+    session: ReplSession,
+    console: Console,
+    *,
+    on_exit: Callable[[], None],
+    confirm_fn: Callable[[str], str] | None = None,
+) -> None:
+    decision = _router.route_input(text, session)
+    kind = decision.route_kind.value
+    if kind in ("follow_up", "new_alert") and _looks_like_correction(text):
+        session.record_intervention("correction")
+    execute_routed_turn(
+        text,
+        session,
+        console,
+        on_exit=on_exit,
+        confirm_fn=confirm_fn,
+        decision=decision,
+    )
+
+
+def run_initial_input(
+    initial_input: str,
+    session: ReplSession,
+    hot_reloader: HotReloadCoordinator | None = None,
+) -> int:
+    console = Console(
+        highlight=False,
+        force_terminal=True,
+        color_system="truecolor",
+        legacy_windows=False,
+    )
+    render_banner(console)
+    exit_requested = [False]
+
+    def _early_exit() -> None:
+        exit_requested[0] = True
+
+    for line in initial_input.splitlines():
+        if hot_reloader is not None:
+            hot_reloader.check_and_reload(console)
+        stripped = line.strip()
+        if not stripped:
+            continue
+        render_submitted_prompt(console, session, stripped)
+        dispatch_one_turn(stripped, session, console, on_exit=_early_exit)
+        if exit_requested[0]:
+            return 0
+    return 0
+
+
+def route_confirm_through_prompt(state: ReplState, prompt_text: str) -> str:
+    sys.stdout.write(prompt_text)
+    sys.stdout.flush()
+
+    response_event = threading.Event()
+    state.confirm_response = []
+    state.confirm_event = response_event
+    try:
+        while not response_event.is_set():
+            cancel = state.current_cancel_event
+            if cancel is not None and cancel.is_set():
+                raise DispatchCancelled("cancelled while awaiting confirmation")
+            response_event.wait(timeout=_PROMPT_REFRESH_INTERVAL_S)
+        if not state.confirm_response:
+            raise DispatchCancelled("cancelled while awaiting confirmation")
+        return state.confirm_response[0]
+    finally:
+        state.confirm_event = None
+        state.confirm_response = []
+
+
+def build_cancel_key_bindings(state: ReplState) -> KeyBindings:
+    kb = KeyBindings()
+
+    @kb.add("escape", eager=True)
+    def _on_escape(event: KeyPressEvent) -> None:
+        if state.is_dispatch_running():
+            state.cancel_current_dispatch()
+            return
+        if event.current_buffer.text:
+            event.current_buffer.reset()
+
+    @kb.add("c-l")
+    def _on_ctrl_l(event: KeyPressEvent) -> None:
+        event.app.renderer.clear()
+
+    return kb
+
+
+def install_session_key_bindings(pt_session: object, extra_kb: KeyBindings) -> None:
+    existing = getattr(pt_session, "key_bindings", None)
+    merged = merge_key_bindings([existing, extra_kb]) if existing is not None else extra_kb
+    pt_session.key_bindings = merged  # type: ignore[attr-defined]
+
+
+# Backward-compatible aliases for existing tests/imports.
+_looks_like_confirmation_answer = looks_like_confirmation_answer
+_looks_like_cancel_request = looks_like_cancel_request
+_dispatch_should_show_spinner = dispatch_should_show_spinner
+_dispatch_needs_exclusive_stdin = dispatch_needs_exclusive_stdin
+_dispatch_one_turn = dispatch_one_turn
+_run_initial_input = run_initial_input
+_route_confirm_through_prompt = route_confirm_through_prompt
+_build_cancel_key_bindings = build_cancel_key_bindings
+_install_session_key_bindings = install_session_key_bindings
+
+__all__ = [
+    "DispatchCancelled",
+    "build_cancel_key_bindings",
+    "dispatch_needs_exclusive_stdin",
+    "dispatch_one_turn",
+    "dispatch_should_show_spinner",
+    "install_session_key_bindings",
+    "looks_like_cancel_request",
+    "looks_like_confirmation_answer",
+    "route_confirm_through_prompt",
+    "run_initial_input",
+    "_build_cancel_key_bindings",
+    "_dispatch_needs_exclusive_stdin",
+    "_dispatch_one_turn",
+    "_dispatch_should_show_spinner",
+    "_install_session_key_bindings",
+    "_looks_like_cancel_request",
+    "_looks_like_confirmation_answer",
+    "_route_confirm_through_prompt",
+    "_run_initial_input",
+]
