+++ app/cli/interactive_shell/prompting/prompt_surface.py
@@ -19,7 +19,7 @@ from rich.text import Text
 
 from app.cli.interactive_shell.commands import SLASH_COMMANDS
 from app.cli.interactive_shell.history import load_prompt_history
-from app.cli.interactive_shell.routing.router import BARE_COMMAND_ALIASES
+from app.cli.interactive_shell.routing.aliases import BARE_COMMAND_ALIASES
 from app.cli.interactive_shell.runtime import ReplSession
 from app.cli.interactive_shell.ui import (
     ANSI_DIM,
