+++ app/cli/interactive_shell/routing/tests/prompts/integration_llm_route_input_cases.yml
@@ -1,0 +1,32 @@
+- id: synthetic_connection_id
+  input: run synthetic test 002-connection-exhaustion
+  llm_route: cli_agent
+  expected_kind: cli_agent
+  expected_signals:
+    - intent_classifier_llm
+
+- id: synthetic_memory_id
+  input: run synthetic test 005-memory-pressure
+  llm_route: cli_agent
+  expected_kind: cli_agent
+
+- id: synthetic_cpu_id
+  input: launch synthetic test 003-cpu-spike
+  llm_route: cli_agent
+  expected_kind: cli_agent
+
+- id: real_alert_text
+  input: the checkout service is returning 502 errors for 30% of requests
+  llm_route: new_alert
+  expected_kind: new_alert
+
+- id: follow_up_with_prior
+  input: what caused the spike?
+  llm_route: follow_up
+  expected_kind: follow_up
+  with_prior_state: true
+
+- id: cli_help_docs_question
+  input: how do I set up the datadog integration?
+  llm_route: cli_help
+  expected_kind: cli_help
