+++ app/cli/interactive_shell/routing/tests/prompts/routing_cases.yml
